NL_ tables fields types alignment - take 2 - not final. It compiles but most probably ain't no works.

+++ Backend/Strategies/MainStrategy/MainStrategy.cs
@@ -898,7 +898,7 @@
 		private int offeredCreditLine;
 
 		private readonly InternalCashRequestID cashRequestID;
-		private int nlCashRequestID;
+		private long nlCashRequestID;
 
 		private readonly StrategiesMailer mailer;
